+++ changelog.md
@@ -1,4 +1,10 @@
 
+## V4.0.8 Fix Bug Context Menu disapear
+
+- Fix Bug Context Menu disapear --> git commit -a -m "V4.0.8 Fix Ctx Menu Editor"
+- Modify "Lien URL/Email" --> "<> Lien URL/Email" for Ctx Menu Editor and Menu + translations
+
+
 ## V4.0.7 Fix Issue [#111] [#110] [#102]
 
 - Fix Issue [#111](https://github.com/lephotographelibre/BlueNotebook/issues/111) V4.0.6 Add OS version to the "About" WIndows #111